requirements.txt
streamlit==1.30.0
flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
langchain~=0.1.12
numpy==1.24.3
pandas==2.0.1
bs4~=0.0.2
python-dotenv~=1.0.1
docx~=0.2.4
fitz~=0.0.1.dev2
faiss